PR target/53912 	* pointer-set.c (hash1): Cast from pointer via uintptr_t.
git-svn-id: svn+ssh://gcc.gnu.org/svn/gcc/trunk@194067 138bc75d-0d04-0410-961f-82ee72b054a4

+++ gcc/pointer-set.c
@@ -64,7 +64,7 @@ hash1 (const void *p, unsigned long max, unsigned long logmax)
 #endif
   const unsigned long shift = HOST_BITS_PER_LONG - logmax;
 
-  return ((A * (unsigned long) p) >> shift) & (max - 1);
+  return ((A * (uintptr_t) p) >> shift) & (max - 1);
 }
 
 /* Allocate an empty pointer set.  */
